🎨 refactor Listings.ts
