+++ src/classes/Listings/details/painted.ts
@@ -1,0 +1,32 @@
+export default function replacePainted(name: string): string {
+    return name
+        .replace(/A Color Similar to Slate/, '🧪')
+        .replace(/A Deep Commitment to Purple/, '🪀')
+        .replace(/A Distinctive Lack of Hue/, '🎩')
+        .replace(/A Mann's Mint/, '👽')
+        .replace(/After Eight/, '🏴')
+        .replace(/Aged Moustache Grey/, '👤')
+        .replace(/An Extraordinary Abundance of Tinge/, '🏐')
+        .replace(/Australium Gold/, '🏆')
+        .replace(/Color No. 216-190-216/, '🧠')
+        .replace(/Dark Salmon Injustice/, '🐚')
+        .replace(/Drably Olive/, '🥝')
+        .replace(/Indubitably Green/, '🥦')
+        .replace(/Mann Co. Orange/, '🏀')
+        .replace(/Muskelmannbraun/, '👜')
+        .replace(/Noble Hatter's Violet/, '🍇')
+        .replace(/Peculiarly Drab Tincture/, '🪑')
+        .replace(/Pink as Hell/, '🎀')
+        .replace(/Radigan Conagher Brown/, '🚪')
+        .replace(/The Bitter Taste of Defeat and Lime/, '💚')
+        .replace(/The Color of a Gentlemann's Business Pants/, '🧽')
+        .replace(/Ye Olde Rustic Colour/, '🥔')
+        .replace(/Zepheniah's Greed/, '🌳')
+        .replace(/An Air of Debonair/, '👜🔷')
+        .replace(/Balaclavas Are Forever/, '👜🔷')
+        .replace(/Operator's Overalls/, '👜🔷')
+        .replace(/Cream Spirit/, '🍘🥮')
+        .replace(/Team Spirit/, '🔵🔴')
+        .replace(/The Value of Teamwork/, '👨🏽‍🤝‍👨🏻')
+        .replace(/Waterlogged Lab Coat/, '👨🏽‍🤝‍👨🏽');
+}
