+++ src/classes/Listings/details/sheens.ts
@@ -1,0 +1,10 @@
+export default function replaceSheens(name: string): string {
+    return name
+        .replace(/Team Shine/, '🔵🔴')
+        .replace(/Hot Rod/, '🎗️')
+        .replace(/Manndarin/, '🟠')
+        .replace(/Deadly Daffodil/, '🟡')
+        .replace(/Mean Green/, '🟢')
+        .replace(/Agonizing Emerald/, '🟩')
+        .replace(/Villainous Violet/, '🟣');
+}
