+++ src/classes/Listings/details/spells.ts
@@ -1,0 +1,19 @@
+export default function replaceSpells(name: string): string {
+    return name
+        .replace(/Putrescent Pigmentation/, 'PP 🍃')
+        .replace(/Die Job/, 'DJ 🍐')
+        .replace(/Chromatic Corruption/, 'CC 🪀')
+        .replace(/Spectral Spectrum/, 'Spec 🔵🔴')
+        .replace(/Sinister Staining/, 'Sin 🍈')
+        .replace(/Voices From Below/, 'VFB 🗣️')
+        .replace(/Team Spirit Footprints/, 'TS-FP 🔵🔴')
+        .replace(/Gangreen Footprints/, 'GG-FP 🟡')
+        .replace(/Corpse Gray Footprints/, 'CG-FP 👽')
+        .replace(/Violent Violet Footprints/, 'VV-FP ♨️')
+        .replace(/Rotten Orange Footprints/, 'RO-FP 🍊')
+        .replace(/Bruised Purple Footprints/, 'BP-FP 🍷')
+        .replace(/Headless Horseshoes/, 'HH 🍇')
+        .replace(/Exorcism/, '👻')
+        .replace(/Pumpkin Bomb/, '🎃💣')
+        .replace(/Halloween Fire/, '🔥🟢');
+}
