+++ src/classes/Listings/export.ts
@@ -1,0 +1,6 @@
+import replaceSpells from './details/spells';
+import replaceKillstreaker from './details/killstreakers';
+import replaceSheens from './details/sheens';
+import replacePainted from './details/painted';
+
+export { replaceSpells, replaceKillstreaker, replaceSheens, replacePainted };
